fix: Overhaul config system to fix critical startup bug and data loss
This commit addresses a critical bug that caused the addon to fail on startup and reset player ranks. This was caused by a circular dependency and shallow copy issues in the configuration manager.

This commit introduces a major overhaul of the configuration system and addresses a wide range of feature requests and bug fixes.

- **Critical Bug Fix:**
  - Resolved a `TypeError: not a function` crash on startup by creating a new `objectUtils.js` file to break a circular dependency between `configManager.js` and `utils.js`.
  - Replaced all shallow copies in the config system with deep copies to prevent data loss and ensure the `/xreload` logic works correctly with nested objects.

- **Config System Overhaul:**
  - Refactored `configManager.js` to manage two persisted configurations: `currentConfig` (for live, in-game state) and `lastLoadedConfig` (a snapshot from the last file load).
  - Implemented a version check and migration system to handle addon updates gracefully, preserving user settings.
  - The addon now automatically performs a config reload on every startup (except the very first one) to apply changes from the `config.js` file.
  - The `ownerPlayerNames` list is now always loaded directly from the file to ensure it remains authoritative.

- **Feature Fixes & Implementations:**
  - Fixed the "Welcomer" and "Death Coords" features.
  - Added a new `/deathcoords` command.
  - Fixed the `/debug` command to correctly toggle its state.
  - Fixed redundant messages in the TPA system.

- **Versioning & Release Workflow:**
  - Standardized the version placeholder to `[1, 0, 0]` across all files for consistency.
  - Updated the release workflow to handle the new versioning system.

- **Command and Alias Updates:**
  - Added a large number of user-approved aliases to various commands.
  - Updated the `commandSettings` in the config to include all commands.

- **Documentation:**
  - Updated all relevant documentation to reflect these new features and changes.
